Rtos issues (#328)
* Fixed issues

* Fixed comment

* Separated Amps task and app.

* Made BSP CAN compatible with RTOS, fixed CAN and SPI IRQ handlers. Also added MinionsASIC_Mutex to tasks.c instead of in its own file

* Fixed interrupt handlers

* Reverted files with unnecessary changes

* Added Fault_BitMap to Task_Amperes.c

Required to finish merging rtos into rtos_issues2
