+++ Apps/Src/Tasks.c
@@ -57,6 +57,7 @@ OS_SEM Fault_Sem4;
  * Mutexes
  */
 OS_MUTEX WDog_Mutex;
+OS_MUTEX MinionsASIC_Mutex;
 
 /**
  * Global Variables
