+++ BSP/STM32F413/Src/BSP_CAN.c
@@ -220,10 +220,12 @@ ErrorStatus BSP_CAN_Read(uint32_t *id, uint8_t *data) {
 }
 
 void CAN1_RX0_IRQHandler(void) {
-    asm("CPSID I");   // Disable Interrupts
-    OSIntEnter();     // Signal to uC/OS
-    asm("CPSIE I");   // Re-enable interrupts
-    
+    #ifdef RTOS
+    CPU_SR_ALLOC();
+    CPU_CRITICAL_ENTER();
+    OSIntEnter();
+    CPU_CRITICAL_EXIT();
+    #endif
     // Take any pending messages into a queue
     while(CAN_MessagePending(CAN1, CAN_FIFO0)) {
         CAN_Receive(CAN1, CAN_FIFO0, &gRxMessage);
@@ -241,20 +243,24 @@ void CAN1_RX0_IRQHandler(void) {
             }
         }
     }
-
+    #ifdef RTOS
     OSIntExit();      // Signal to uC/OS
+    #endif
 }
 
 void CAN1_TX_IRQHandler(void) {
-    asm("CPSID I");   // Disable Interrupts
-    OSIntEnter();     // Signal to uC/OS
-    asm("CPSIE I");   // Re-enable interrupts
-
+    #ifdef RTOS
+    CPU_SR_ALLOC();
+    CPU_CRITICAL_ENTER();
+    OSIntEnter();
+    CPU_CRITICAL_EXIT();
+    #endif
     // Acknowledge 
     CAN_ClearFlag(CAN1, CAN_FLAG_RQCP0 | CAN_FLAG_RQCP1 | CAN_FLAG_RQCP2);
 
     // Call the function provided
     gTxEnd();
-
+    #ifdef RTOS
     OSIntExit();      // Signal to uC/OS
+    #endif
 }
